requirements.txt
discord.py>=2.3.0
google-genai>=1.0.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
redis>=5.0.0
alembic>=1.13.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
